refactor: objectUtils 유틸 분리
